feat: Enhance home page with course retrieval and display features
- Updated FragmentController to include methods for fetching top courses and paginated course lists.
- Introduced CourseDTO for better data handling and updated Course entity to use BigDecimal for price and discount.
- Enhanced Thymeleaf templates to dynamically display top courses and paginated course lists.
- Improved CourseMapper to support conversion between Course and CourseDTO.
- Added new methods in CourseService and CourseServiceImpl for retrieving top courses and courses sorted by ratings.

+++ src/main/java/com/OLearning/dto/CourseDTO.java
@@ -1,12 +1,12 @@
 package com.OLearning.dto;
 
-import com.OLearning.entity.Category;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -17,11 +17,10 @@ import java.time.LocalDateTime;
 public class CourseDTO {
 
     private Long courseId;
-
     private String title;
     private String description;
-    private Double price;
-    private Double discount;
+    private BigDecimal price;
+    private BigDecimal discount;
     private String courseImg;
     private Integer duration;
     private Integer totalLessons;
