+++ src/main/java/com/OLearning/mapper/CourseDetailMapper.java
@@ -14,8 +14,8 @@ public class CourseDetailMapper {
         dto.setCourseId(course.getCourseId());
         dto.setTitle(course.getTitle());
         dto.setDescription(course.getDescription());
-        dto.setPrice(course.getPrice());
-        dto.setDiscount(course.getDiscount());
+        dto.setPrice(course.getPrice().doubleValue());
+        dto.setDiscount(course.getDiscount().doubleValue());
         dto.setCourseImg(course.getCourseImg());
         dto.setDuration(course.getDuration());
         dto.setTotalLessons(course.getTotalLessons());
